Relocate endpoints functions to controllers directory
